Update sqrt_helper function signature

+++ 0x08-recursion/main.h
@@ -8,6 +8,6 @@ int _strlen_recursion(char *s);
 int factorial(int n);
 int _pow_recursion(int x, int y);
 int _sqrt_recursion(int n);
-long int sqrt_helper(int n, int start, int end);
+long int sqrt_helper(long int n, long int start, long int end);
 
 #endif
